Extract path related logic from create command
